+++ plugins/inputs/filecount/filesystem_helpers_test.go
@@ -84,8 +84,8 @@ func getTestFileSystem() fakeFileSystem {
 	dmask |= (1 << uint(32-1))
 
 	fileList := map[string]fakeFileInfo{
-		"/testdata":     {name: "testdata", size: int64(4096), filemode: uint32(dmask), modtime: mtime, isdir: true},
-		"/testdata/foo": {name: "foo", filemode: uint32(fmask), modtime: mtime},
+		"/testdata":     {name: "testdata", size: int64(4096), filemode: dmask, modtime: mtime, isdir: true},
+		"/testdata/foo": {name: "foo", filemode: fmask, modtime: mtime},
 	}
 
 	fs := fakeFileSystem{files: fileList}
