+++ plugins/inputs/mongodb/mongostat.go
@@ -1181,7 +1181,7 @@ func NewStatLine(oldMongo, newMongo MongoStatus, key string, all bool, sampleSec
 				// Get the entry with the highest lock
 				highestLocked := lockdiffs[len(lockdiffs)-1]
 
-				timeDiffMillis := int64(newStat.UptimeMillis - oldStat.UptimeMillis)
+				timeDiffMillis := newStat.UptimeMillis - oldStat.UptimeMillis
 
 				lockToReport := highestLocked.Writes
